refactor: consolidate OutputFormat enum into commands/mod.rs
Move the shared OutputFormat enum (Text/Json) from up/mod.rs to
commands/mod.rs and remove the duplicate definition from build.rs.
All commands now import from a single canonical location. Normalize
value_enum attribute on prune's --output arg for consistency.
