+++ crates/cella-cli/src/commands/nvim.rs
@@ -4,7 +4,8 @@ use tracing::debug;
 
 use cella_backend::{ExecOptions, InteractiveExecOptions};
 
-use super::up::{OutputFormat, UpArgs, UpContext};
+use super::OutputFormat;
+use super::up::{UpArgs, UpContext};
 
 use crate::picker;
